+++ pkg/types/resolver_error.go
@@ -3,26 +3,28 @@ package types
 import "errors"
 
 const (
-	ErrorServiceTypeNotRegistered              = "service type is not registered"
-	ErrorRequiredServiceNotRegistered          = "required service type is not registered"
-	ErrorCannotResolveService                  = "cannot resolve service"
-	ErrorAmbiguousServiceInstancesResolved     = "the resolve operation resulted in multiple service instances"
-	ErrorActivatorFunctionInvalidReturnType    = "activator function has an invalid return type"
-	ErrorCircularDependencyDetected            = "circular dependency detected"
-	ErrorCannotBuildDependencyGraph            = "failed to build dependency graph"
-	ErrorInstanceCannotBeNil                   = "instance cannot be nil"
-	ErrorServiceTypeMustBeInterface            = "service type must be an interface"
-	ErrorCannotRegisterTypeWithResolverOptions = "cannot register type with resolver options"
+	ErrorServiceTypeNotRegistered               = "service type is not registered"
+	ErrorRequiredServiceNotRegistered           = "required service type is not registered"
+	ErrorCannotResolveService                   = "cannot resolve service"
+	ErrorAmbiguousServiceInstancesResolved      = "the resolve operation resulted in multiple service instances"
+	ErrorActivatorFunctionInvalidReturnType     = "activator function has an invalid return type"
+	ErrorCircularDependencyDetected             = "circular dependency detected"
+	ErrorCannotBuildDependencyGraph             = "failed to build dependency graph"
+	ErrorInstanceCannotBeNil                    = "instance cannot be nil"
+	ErrorServiceTypeMustBeInterface             = "service type must be an interface"
+	ErrorCannotRegisterTypeWithResolverOptions  = "cannot register type with resolver options"
+	ErrorCannotCreateInstanceOfUnregisteredType = "failed to create instance of unregistered type"
 )
 
 var (
-	ErrServiceTypeNotRegistered              = errors.New(ErrorServiceTypeNotRegistered)
-	ErrActivatorFunctionInvalidReturnType    = errors.New(ErrorCannotResolveService)
-	ErrCannotBuildDependencyGraph            = errors.New(ErrorCannotBuildDependencyGraph)
-	ErrCircularDependencyDetected            = errors.New(ErrorCircularDependencyDetected)
-	ErrInstanceCannotBeNil                   = errors.New(ErrorInstanceCannotBeNil)
-	ErrServiceTypeMustBeInterface            = errors.New(ErrorServiceTypeMustBeInterface)
-	ErrCannotRegisterTypeWithResolverOptions = errors.New(ErrorCannotRegisterTypeWithResolverOptions)
+	ErrServiceTypeNotRegistered               = errors.New(ErrorServiceTypeNotRegistered)
+	ErrActivatorFunctionInvalidReturnType     = errors.New(ErrorCannotResolveService)
+	ErrCannotBuildDependencyGraph             = errors.New(ErrorCannotBuildDependencyGraph)
+	ErrCircularDependencyDetected             = errors.New(ErrorCircularDependencyDetected)
+	ErrInstanceCannotBeNil                    = errors.New(ErrorInstanceCannotBeNil)
+	ErrServiceTypeMustBeInterface             = errors.New(ErrorServiceTypeMustBeInterface)
+	ErrCannotRegisterTypeWithResolverOptions  = errors.New(ErrorCannotRegisterTypeWithResolverOptions)
+	ErrCannotCreateInstanceOfUnregisteredType = errors.New(ErrorCannotCreateInstanceOfUnregisteredType)
 )
 
 type ResolverError struct {